Cleaning Entitys

+++ src/main/java/dev/jabivullah/parking_lot_management_system/entity/Operator.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 public class Operator extends BaseEntity{
     private String name;
     private String phoneNumber;
-    private int empId;
+    private Long empId;
 }
